Updated some text and updated the /view/ function

+++ include/Login.php
@@ -18,7 +18,8 @@ class Login
         }
     }
 
-    public function load() {
+    public function load()
+    {
         $this->Database = new Database();
     }
 
@@ -38,8 +39,8 @@ class Login
                     $token = uniqid('', true);
                     $hashToken = password_hash($token, PASSWORD_BCRYPT);
 
-                    setcookie('token', $token, time()+60*60*24*30, '/', 'scoredb.julianvos.nl', false, true);
-                    setcookie('username', $username, time()+60*60*24*30, '/', 'scoredb.julianvos.nl', false, true);
+                    setcookie('token', $token, time()+60*60*24*30, '/', $this->Config->getEnv(), false, true);
+                    setcookie('username', $username, time()+60*60*24*30, '/', $this->Config->getEnv(), false, true);
 
                     echo $username . $token;
 
@@ -66,44 +67,50 @@ class Login
 
     public function check($username, $token)
     {
-        if ($stmt = $this->con->prepare('SELECT token FROM accounts WHERE username = ?')) {
-            $stmt->bind_param('s', $username);
-            $stmt->execute();
-            $stmt->store_result();
+        if ($username != '' && $token != '') {
+            if ($stmt = $this->con->prepare('SELECT token FROM accounts WHERE username = ?')) {
+                $stmt->bind_param('s', $username);
+                $stmt->execute();
+                $stmt->store_result();
     
-            if ($stmt->num_rows > 0) {
-                $stmt->bind_result($tokenDB);
-                $stmt->fetch();
+                if ($stmt->num_rows > 0) {
+                    $stmt->bind_result($tokenDB);
+                    $stmt->fetch();
         
-                if (password_verify($token, $tokenDB)) {
-                    return true;
+                    if (password_verify($token, $tokenDB)) {
+                        return true;
+                    } else {
+                        echo '<p>Could not verify user. You will be logged out.';
+                        sleep(5);
+                        $this->log_out();
+                        return false;
+                    }
                 } else {
-                    echo '<p>Could not verify user. You will be logged out in 5 seconds.';
+                    echo '<p>Could not verify user. You will be logged out.';
                     sleep(5);
                     $this->log_out();
                     return false;
                 }
+                $stmt->close();
             } else {
-                echo '<p>Could not verify user. You will be logged out in 5 seconds.';
+                echo '<p>Could not verify user. You will be logged out.';
                 sleep(5);
                 $this->log_out();
                 return false;
             }
-            $stmt->close();
-        } else {
-            echo '<p>Could not verify user. You will be logged out in 5 seconds.';
-            sleep(5);
-            $this->log_out();
-            return false;
         }
     }
 
     public function log_out()
     {
-        $this->load();
-        setcookie('token', '', time() - 3600);
-        setcookie('username', '', time() - 3600);
-        $this->Database->deleteToken($_COOKIE['username']);
-        header('location: /index.php');
+        if (isset($_COOKIE['username']) && isset($_COOKIE['token'])) {
+            $this->load();
+            $this->Database->deleteToken($_COOKIE['username']);
+            unset($_COOKIE['username']);
+            unset($_COOKIE['token']);
+            setcookie('token', '', time()-3600, '/', $this->Config->getEnv(), false, true);
+            setcookie('username', '', time()-3600, '/', $this->Config->getEnv(), false, true);
+            echo '<p>You have been logged out. <a href="../index.php">Return to home</a></p>';
+        }
     }
 }
